ref: Move language_parsers to source_code_management
