+++ src/sentry/integrations/github/constants.py
@@ -1,6 +1,3 @@
 ISSUE_LOCKED_ERROR_MESSAGE = "Unable to create comment because issue is locked."
 
 RATE_LIMITED_MESSAGE = "API rate limit exceeded"
-
-# Number of stackframes to check for filename + function combo, starting from the top
-STACKFRAME_COUNT = 4
