+++ src/sentry/integrations/source_code_management/constants.py
@@ -1,0 +1,2 @@
+# Number of stackframes to check for filename + function combo, starting from the top
+STACKFRAME_COUNT = 4
