+++ tests/sentry/integrations/github/tasks/test_open_pr_comment.py
@@ -6,8 +6,6 @@ import responses
 from django.utils import timezone
 
 from sentry.constants import ObjectStatus
-from sentry.integrations.github.constants import STACKFRAME_COUNT
-from sentry.integrations.github.tasks.language_parsers import PATCH_PARSERS
 from sentry.integrations.github.tasks.open_pr_comment import (
     format_issue_table,
     format_open_pr_comment,
@@ -20,6 +18,8 @@ from sentry.integrations.github.tasks.open_pr_comment import (
 )
 from sentry.integrations.github.tasks.utils import PullRequestFile, PullRequestIssue
 from sentry.integrations.models.integration import Integration
+from sentry.integrations.source_code_management.constants import STACKFRAME_COUNT
+from sentry.integrations.source_code_management.language_parsers import PATCH_PARSERS
 from sentry.models.group import Group, GroupStatus
 from sentry.models.pullrequest import CommentType, PullRequest, PullRequestComment
 from sentry.shared_integrations.exceptions import ApiError
@@ -825,7 +825,7 @@ Your pull request is modifying functions with the following pre-existing issues:
     "sentry.integrations.github.tasks.open_pr_comment.get_projects_and_filenames_from_source_file"
 )
 @patch(
-    "sentry.integrations.github.tasks.language_parsers.PythonParser.extract_functions_from_patch"
+    "sentry.integrations.source_code_management.language_parsers.PythonParser.extract_functions_from_patch"
 )
 @patch("sentry.integrations.github.tasks.open_pr_comment.get_top_5_issues_by_count_for_file")
 @patch("sentry.integrations.github.tasks.open_pr_comment.safe_for_comment")
